refactor(adr-018): rewire runThreeSessionTdd to use runTddSessionOp x3; move ThreeSessionTddOptions to types.ts
- Move ThreeSessionTddOptions interface from orchestrator.ts to types.ts (with all JSDoc)
- Move sumTddTokenUsage function from orchestrator.ts to types.ts and export it
- Rewire Session 1/2/3 in runThreeSessionTdd to use runTddSessionOp instead of runTddSession directly
- Remove buildInteractionBridge, InteractionChain imports from orchestrator.ts (now internal to session-op.ts)
- Remove runTddSession from orchestrator.ts imports (kept rollbackToRef, truncateTestOutput)
- Update orchestrator-ctx.ts to import ThreeSessionTddOptions from ./types
- Add ThreeSessionTddOptions to src/tdd/index.ts barrel export from ./types
- Re-export ThreeSessionTddOptions from orchestrator.ts for backward compat
- orchestrator.ts reduced from 535 to 397 lines (under 400-line limit)
